Add multiboot info structure parsing

+++ kernel/modules.c
@@ -2,7 +2,7 @@
 #include <tinyos/kernel/panic.h>
 
 static const struct module *modules[64];
-static unsigned module_idx = 0;
+static size_t module_idx = 0;
 
 void register_module(const struct module *m) {
   assert(module_idx < ARRAY_SIZE(modules));
@@ -11,8 +11,7 @@ void register_module(const struct module *m) {
 
 void init_modules(void) {
   /* TODO: sort modules by their priority */
-
-  for (unsigned i = 0; i < module_idx; i++) {
+  for (size_t i = 0; i < module_idx; i++) {
     const struct module *m = modules[i];
     m->init();
   }
